emulate uninstall
uninstall use case: take out a device or alias that is not desired.
example: gnucsator. Explicitly removing some bits after loading default
plugins. (may be seen as a workaround, but that's a different story).

Public uninstall has gone. can be achieved (approximated) by
INSTALL'ing a NULL pointer. But there is an assertion in the way.
Without it, everything seems fine.