logic-4: swap PWR_NODE (vdd) and GND_NODE (vss) pins on logic devices 	change load_passive load_point.
In next commit, plan to eliminate 3 pins,
to make pinout compliant with Verilog-HDL LRM.
but this caused problem because other pins were referenced to GND_NODE pin,
which was convenient, but doesn't work with LRM pinout.
So this is an interim change to put a DC voltage there to show impact.