event-1: do not quantize on rejection.
It has no benefit.
This changes step control at a rejection, which has a trivial
effect on many regressions.

On old code, it tried to choose a step that would lead to a long repeat
to a smooth ending at a fixed_time (was the rejected step, which has
no significance) but there are so many steps to get there, with
variation, such that the basis was false.