event-2: cleanup.
The step control is mostly integer-like, but with type double in the class.
Going in, it is scaled by dtmin and rounded, so it is mostly integer-like,
but retaining the overflow features of floating point.

Arithmetic operations are not rounded again, so it can go off-grid on
repeated equal time steps.  This is preferable over changing to make
the ends come out.