final-step-2: ac and fourier, and some other nits.
This appears to work correctly and passes all regressions,
but it really doesn't.
??_final should not be called if it doesn't really run,
for example when an exception is thrown during elaboration.
This leads to question whether it should be called when
a sweep runs partially but does not complete.
Previously I said it should, but I am not sure about that.
In any case, it definitely should not run when elaboration
does not complete or in PRESET mode.