final_step-2: separate final and finish.
"finish" is cleanup, always done even if exception is called,
even if during elaboration, even if no simulation actually runs.

"final" is post-processing, done if any sim is done, even if partial,
but not done if no sim is done, such as when exception thrown
in elaboration.

Aded try block around sweep to do this.