logic-bug:  fix new_event quantization.
Some regressions changed.  Believed correct now, was incorrect before.
The bug was in event quantization.
It sometimes caused events to be scheduled early, so the scheduled time
was sooner than the correct time, by dtmin, which caused nothing to happen
at the queued time, but rather on the next step, resulting in delayed
propagation of change in logic state.

It appears the bug was introduced in 526c1d03, Sep 9,
with quantization, missing _eq::new_event.