logic-bug: Change s-fo.cc to take the max dtmin based on dtratio instead of min.
This is only for Fourier command.  Tran already did max, correctly.
This line adjusts dtmin upward (larger steps) when dtmax is large,
keeping the ratio to dtratio.  This is only for using the default
value of dtmin, from opt, which is a good value for the usual
step sizes, but too small for low frequency runs, which are
overrepresented in the test suite.

There are minor changes in regressions, in noise levels, in a few cases.