bugfix-20241211: fix bug in lib/ap_convert.
	revert apps/lang_verilog from 4ac5ae438c3a640f6b4eed3120451fd971b14058,
	Change no longer needed after fixing lib/ap_convert.

Fix was to remove an incorrect skip(), which caused it to miss
quotes immediately following an embedded use of a begin_quote
character as a non-quote, so it would mess up on things like (foo()),
but would be fine with (foo( )).

This backs out the change to lang_verilog to parse a different way,
not using ctos, which is correct now that ctos is fixed, which reverts
whitespace changes in regressions where either was technically correct.

Still, the question is raised about parsing the "value" part of a parameter
in an arglist.  It could be argued that a full parse might be the better way,
but this was parse,dump,parse-again, which is less efficient, and
also left several other similar cases in lang* unchanged, even though there
are similar issues.  It may make sense to change to parsing the expression
on first pass always, and a new set_param_by_* that takes an Expression as
its argument instead of std::string. but if we do it should be done
everywhere consistently.